+++ dflib/src/main/java/com/nhl/dflib/GroupBy.java
@@ -1,5 +1,6 @@
 package com.nhl.dflib;
 
+import com.nhl.dflib.aggregate.ColumnAggregator;
 import com.nhl.dflib.concat.SeriesConcat;
 import com.nhl.dflib.concat.VConcat;
 import com.nhl.dflib.row.RowProxy;