+++ dflib/src/main/java/com/nhl/dflib/aggregate/AggregatorFunctions.java
@@ -1,26 +1,51 @@
 package com.nhl.dflib.aggregate;
 
-import com.nhl.dflib.RowToValueMapper;
+import com.nhl.dflib.ValueMapper;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Set;
+import java.util.function.BiConsumer;
+import java.util.function.BinaryOperator;
 import java.util.function.Function;
+import java.util.function.Supplier;
 import java.util.stream.Collector;
 
 public class AggregatorFunctions {
 
-    public static RowToValueMapper<String> toString(String column) {
-        RowToValueMapper<?> reader = RowToValueMapper.columnReader(column);
-        return r -> String.valueOf(reader.map(r));
+    public static <S, M, A, T> Collector<S, A, T> mappedCollector(Collector<M, A, T> collector, ValueMapper<S, M> mapper) {
+
+        return new Collector<S, A, T>() {
+            @Override
+            public Supplier<A> supplier() {
+                return collector.supplier();
+            }
+
+            @Override
+            public BiConsumer<A, S> accumulator() {
+                BiConsumer<A, M> accum = collector.accumulator();
+                return (a, s) -> accum.accept(a, mapper.map(s));
+            }
+
+            @Override
+            public BinaryOperator<A> combiner() {
+                return collector.combiner();
+            }
+
+            @Override
+            public Function<A, T> finisher() {
+                return collector.finisher();
+            }
+
+            @Override
+            public Set<Characteristics> characteristics() {
+                return collector.characteristics();
+            }
+        };
     }
 
-    public static RowToValueMapper<String> toString(int column) {
-        RowToValueMapper<?> reader = RowToValueMapper.columnReader(column);
-        return r -> String.valueOf(reader.map(r));
-    }
-
-    public static Collector<Number, List<Double>, Double> medianCollector() {
+    public static <S extends Number> Collector<S, List<Double>, Double> medianCollector() {
         return new AggregationCollector<>(
                 ArrayList::new,
                 (list, d) -> {