+++ dflib/src/main/java/com/nhl/dflib/map/ColumnLocator.java
@@ -1,0 +1,17 @@
+package com.nhl.dflib.map;
+
+import com.nhl.dflib.Index;
+
+@FunctionalInterface
+public interface ColumnLocator {
+
+    static ColumnLocator forPosition(int position) {
+        return i -> position;
+    }
+
+    static ColumnLocator forName(String name) {
+        return i -> i.position(name);
+    }
+
+    int position(Index i);
+}